fix: fixed foreign key reference under plugin config (#1597)
* fix: fixed foreign key reference under plugin config
This change updates gdr version to v1.22.4
Here, we have relaxed the restrictions
added in an earlier deck version to error
out in case of foreign keys under a plugin
config. Now, errors come up only if same
type of entity is nested under an entity.
This change has also fixed the resolution
of foreign keys in case names are used,
instead of IDs.

* chore: fixed runWhen condition in test